api/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
mysql-connector-python
flasgger=0.9.7.1